FreeBSD amd64 fixes.   In the import and export modules transcode through a mangling   include via "import_def.h" resp "export_def.h" declared symbols   static as well extern. Although in the same file, it is   appearantly illegal on this architecture and transcode recieved a   SIGBUS. Now the order is changed, first the static declaration,   than the inclusion if *_def.h. Thanks a lot to Hendrik Scholz for   providing a nice (and fast!) testbox.

+++ import/decode_dv.c
@@ -45,17 +45,17 @@ static int verbose=TC_QUIET;
 
 static unsigned char *bufalloc(size_t size)
 {
-   int buffer_align=getpagesize();
+   long buffer_align=getpagesize();
  
    unsigned char *buf = malloc(size + buffer_align);
 
-   int adjust;
+   long adjust;
 
    if (buf == NULL) {
        fprintf(stderr, "(%s) out of memory", __FILE__);
    }
    
-   adjust = buffer_align - ((int) buf) % buffer_align;
+   adjust = buffer_align - ((long) buf) % buffer_align;
 
    if (adjust == buffer_align)
       adjust = 0;
